% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_trimmed_weights.R
\name{add_trimmed_weights_y2}
\alias{add_trimmed_weights_y2}
\title{Add weights to your data}
\usage{
add_trimmed_weights_y2(
  dataset,
  raked_weights,
  upper_limit = 5,
  lower_limit = 0.05,
  trimmed_weights_name
)
}
\arguments{
\item{dataset}{A dataframe to be weighted}

\item{raked_weights}{Raked weights output}

\item{upper_limit}{DEFAULT = 5, set an upper trim limit for your weights}

\item{lower_limit}{DEFAULT = .05, set a lower trim limit for your weights}

\item{trimmed_weights_name}{An unquoted name to give to your new trimmed weights variable}
}
\value{
A tibble of with weights
}
\description{
Use add_weights_y() to add a weight column to your data by taking the raked weights output from rake_y2() and passing it to this function.
}
\examples{
\dontrun{
municipal_data \%>\% 
add_trimmed_weights_y2(
raked_weights
)
}
}
\keyword{freqs}
\keyword{parameters}
\keyword{population}
\keyword{weights}
