% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_weights_y2.R
\name{add_weights_y2}
\alias{add_weights_y2}
\title{Add weights to your data}
\usage{
add_weights_y2(dataset, raked_weights, weights_name)
}
\arguments{
\item{dataset}{A dataframe to be weighted}

\item{raked_weights}{Raked weights output}

\item{weights_name}{An unquoted name to give to your new weights variable}
}
\value{
A tibble of with weights
}
\description{
Use add_weights_y() to add a weight column to your data by taking the raked weights output from rake_y2() and passing it to this function.
}
\examples{
\dontrun{
municipal_data \%>\% 
add_weights_y2(
raked_weights
)
}
}
\keyword{freqs}
\keyword{parameters}
\keyword{population}
\keyword{weights}
