% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prop_table_y2.R
\name{create_prop_table_y2}
\alias{create_prop_table_y2}
\title{Create a table of population parameters to be used for weighting data}
\usage{
create_prop_table_y2(
  dataset,
  variable,
  population_params,
  add_missing_row = FALSE
)
}
\arguments{
\item{dataset}{A dataframe to be weighted}

\item{variable}{The variable for which you will be supplying population parameters}

\item{population_params}{A named vector containing population parameters}

\item{add_missing_row}{DEFAULT = FALSE, Set to TRUE to add a row to the prop table that takes into account "MISSING" values in the variable for which you are supplying population parameters}
}
\value{
A tibble of with population parameters to the environment
}
\description{
Use create_prop_table_y2() to create nicely formatted population parameter tables to be used for weighting data. To avoid typos, create_prop_table_y2() will automatically create an object in your R environment using the provided variable to create the name. This function is designed to work within the y2weights framework.
}
\examples{
municipal_data \%>\% 
create_prop_table_y2(
s_sex,
c(
'Male' = .49,
'Female' = .5,
'In another way, specify if you wish' = .01
)
)
}
\keyword{freqs}
\keyword{parameters}
\keyword{population}
\keyword{weights}
