% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_target_y2.R
\name{define_target_y2}
\alias{define_target_y2}
\title{Create a table of target population parameters to be used for weighting data}
\usage{
define_target_y2(dataset, variable, target_params)
}
\arguments{
\item{dataset}{A dataframe with variables to be used in weighting.}

\item{variable}{The variable for which you will be supplying population parameters.}

\item{target_params}{A named vector containing target parameters you wish your data to be weighted to. If the variable has a "MISSING" level, you do not need to include it in the target_params, it will be calculated automatically (essentially set to weight those respondents as 1, the mean).}
}
\value{
A tibble with population parameters to the environment
}
\description{
Use define_target_y2() to create a table of target population parameters to be used for weighting data. To avoid typos, define_target_y2() will automatically create an object in your R environment using the provided variable to create the name (e.g., "target_varname").
}
\examples{
municipal_data \%>\%
  define_target_y2(
    s_sex,
    c(
      '1' = .49,
      '2' = .5,
      '3' = .01
    )
  )
}
\keyword{freqs}
\keyword{parameters}
\keyword{population}
\keyword{weights}
