% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_weights_y2.R
\name{evaluate_weights}
\alias{evaluate_weights}
\title{Evaluate the weights created for your data}
\usage{
evaluate_weights(
  dataset,
  ...,
  weight_var,
  nas = TRUE,
  large_miss_limit = 0.05,
  large_movement_limit = 0.1
)
}
\arguments{
\item{dataset}{A dataframe to be weighted}

\item{...}{List of variables to check}

\item{weight_var}{Weight variable}

\item{nas}{DEFAULT = TRUE, remove NA's from calculations}

\item{large_miss_limit}{DEFAULT = .05, set to different level to quickly see large discrepancies between weighted frequencies and target frequencies}

\item{large_movement_limit}{DEFAULT = .1, set to different level to quickly see where weighting caused large shifts from the unweighted data}
}
\value{
A tibble of compared weighting schema
}
\description{
Use add_weights_y() to add a weight column to your data by taking the raked weights output from rake_y2() and passing it to this function.
}
\examples{
\dontrun{
municipal_data \%>\% 
evaluate_weights(
w_sex,
w_race,
weight_var = weights
)
}
}
\keyword{freqs}
\keyword{parameters}
\keyword{population}
\keyword{weights}
