% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_weights_y2.R
\name{evaluate_weights_y2}
\alias{evaluate_weights_y2}
\title{Compare weighted and unweighted data vs the ideal}
\usage{
evaluate_weights_y2(
  dataset,
  ...,
  weight_var,
  nas = TRUE,
  remove_missing = FALSE
)
}
\arguments{
\item{dataset}{A dataframe to be weighted}

\item{...}{List of variables to check}

\item{weight_var}{Weight variable}

\item{nas}{DEFAULT = TRUE, Remove NAs from calculations}

\item{remove_missing}{DEFAULT = FALSE, Before running frequencies, filter out responses that are "MISSING"}
}
\value{
A tibble with the following columns: variable, label, result_unweighted, target, result_weighted, movement, diff_from_target
}
\description{
Use evaluate_weights_y2() to compare given variables on: 1) unweighted data, 2) weighted data, 3) target weighting goals, 4) movement after weighting, and 5) difference from the target.
}
\examples{
municipal_data \%>\%
  define_target_y2(
    s_sex,
    c(
      '1' = .49,
      '2' = .5,
      '3' = .01
    )
  )
  
weights_schema <- municipal_data \%>\% 
  rake_y2(
    s_sex
  )
  
municipal_data$trimmed_weights <- trim_weights_y2(weights_schema)
  
municipal_data \%>\% 
  evaluate_weights_y2(
    s_sex,
    weight_var = trimmed_weights
  )
}
\keyword{freqs}
\keyword{parameters}
\keyword{target}
\keyword{weights}
