% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipal_data}
\alias{municipal_data}
\title{Example municipal data}
\format{
A data frame with 100 observations and 59 variables, including a weights variable
}
\usage{
municipal_data
}
\description{
Pre formatted data similar to what you may see in Y2 municipal surveys
}
\keyword{datasets}
