% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake_y2.R
\name{rake_y2}
\alias{rake_y2}
\title{Rake weights using target parameter tables}
\usage{
rake_y2(dataset, ...)
}
\arguments{
\item{dataset}{A dataframe to be used in creating the survey design.}

\item{...}{List of variables}
}
\value{
A complex, raked, survey.design object. This output will be used in the weights_schema argument of trim_weights_y2().
}
\description{
Use rake_y2() in conjunction with define_target_y2() to rake weights cleanly and with easier to understand errors. rake_y2() will automatically pull target population parameter tables from your environment that have matching names to the variables provided.
}
\examples{
municipal_data \%>\%
  define_target_y2(
    s_sex,
    c(
      '1' = .49,
      '2' = .5,
      '3' = .01
    )
  )
weights_schema <- municipal_data \%>\% 
  rake_y2(
    s_sex
  )
}
\keyword{freqs}
\keyword{parameters}
\keyword{population}
\keyword{weights}
