% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_weights_y2.R
\name{trim_weights_y2}
\alias{trim_weights_y2}
\title{Create a trimmed weights vector}
\usage{
trim_weights_y2(
  svy_design,
  limit_method = c("percentile", "standard deviations", "decimal"),
  upper_limit = 0.95,
  lower_limit = 0.05,
  strict = TRUE
)
}
\arguments{
\item{svy_design}{A complex raked survey design object (can be created using the rake_y2() function).}

\item{limit_method}{(default = 'percentile') The type of limits that will be used for the upper_limit/lower_limit arguments. Currently accepts 'percentile', 'standard deviations', or 'decimal'.}

\item{upper_limit}{(default = 0.95) Set an upper trim limit for your weights (set to the 95th percentile of the untrimmed weights by default). If limit_method is set to 'standard deviations', limit will be the number of standard deviations above the mean to which the weight will be trimmed. If limit_method is set to 'decimal', limit will be the upper numerical limit of the weight.}

\item{lower_limit}{(default = 0.05) Set a lower trim limit for your weights (set to the 5th percentile of the untrimmed weights by default). If limit_method is set to 'standard deviations', limit will be the number of standard deviations below the mean to which the weight will be trimmed. If limit_method is set to 'decimal', limit will be the lower numerical limit of the weight.}

\item{strict}{(default = TRUE) The reapportionment of the ‘trimmings’ from the weights may sometimes push other weights over the limits. If strict = TRUE the function repeats the trimming iteratively to prevent this.}
}
\value{
A vector of trimmed weights constructed using the svy_design object
}
\description{
Use trim_weights_y2() to create a vector of trimmed weights by taking the survey design output from rake_y2() and passing it to this function.
}
\examples{
municipal_data \%>\%
  define_target_y2(
    s_sex,
    c(
      '1' = .49,
      '2' = .5,
      '3' = .01
    )
  )
svy_design <- municipal_data \%>\% 
  rake_y2(
    s_sex
  )
  
#' # Default method (percentile)
municipal_data \%>\% 
  dplyr::mutate(
    trimmed_weights = trim_weights_y2(svy_design)
  )

# Standard deviations method
municipal_data \%>\% 
  dplyr::mutate(
    trimmed_weights = trim_weights_y2(
      svy_design,
      limit_method = 'standard deviations',
      upper_limit = 3,
      lower_limit = -3
    )
  )

# Standard deviations method
municipal_data \%>\% 
  dplyr::mutate(
    trimmed_weights = trim_weights_y2(
      svy_design,
      limit_method = 'decimal',
      upper_limit = 5,
      lower_limit = 0.5
    )
  )
}
\keyword{freqs}
\keyword{parameters}
\keyword{population}
\keyword{weights}
