% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_weights_y2.R
\name{evaluate_weights_y2}
\alias{evaluate_weights_y2}
\title{Compare weighted and unweighted data vs the ideal}
\usage{
evaluate_weights_y2(
  dataset,
  ...,
  weight_var,
  nas = TRUE,
  remove_missing = FALSE
)
}
\arguments{
\item{dataset}{A dataframe to be weighted}

\item{...}{List of variables to check}

\item{weight_var}{Weight variable}

\item{nas}{DEFAULT = TRUE, Remove NA's from calculations}

\item{remove_missing}{DEFAULT = FALSE, Before running frequencies, filter out responses that are "MISSING"}
}
\value{
A tibble of target population parameters
}
\description{
Use add_weights_y() to add a weight column to your data by taking the raked weights output from rake_y2() and passing it to this function.
}
\examples{
municipal_data \%>\%
  define_target_y2(
    s_sex,
    c(
      '1' = .49,
      '2' = .5,
      '3' = .01
    )
  )
  
weights_schema <- municipal_data \%>\% 
  rake_y2(
    s_sex
  )
  
municipal_data$trimmed_weights <- trim_weights_y2(weights_schema)
  
municipal_data \%>\% 
  evaluate_weights_y2(
    s_sex,
    weight_var = trimmed_weights
  )
}
\keyword{freqs}
\keyword{parameters}
\keyword{population}
\keyword{weights}
