% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake_y2.R
\name{rake_y2}
\alias{rake_y2}
\title{Rake weights using population parameter tables}
\usage{
rake_y2(dataset, ...)
}
\arguments{
\item{dataset}{A dataframe to be weighted}

\item{...}{List of variables}
}
\value{
A complex survey object
}
\description{
Use rake_y2() in conjunction with create_prop_table_y2() to rake weights cleanly and with easier to understand errors. rake_y2() will automatically pull population parameter tables
}
\examples{
\dontrun{
survey::svydesign(
ids = ~1, 
data = municipal_data
)
}
}
\keyword{freqs}
\keyword{parameters}
\keyword{population}
\keyword{weights}
